perf(leds): use integers instead of f32 for time
Changing Transition.duration to u16 (to save space) and
doing interpolation calculations on fixed-point u32 numbers.
This yields performance improvement of keyboard.update_leds
in the order of 1.981 ms -> 785 us.

+++ src/keyboard/leds/mod.rs
@@ -155,8 +155,11 @@ pub enum Repeat {
 /// Single color transition in a pattern
 #[derive(Debug, PartialEq)]
 pub struct Transition {
+    /// "Destination" color
     pub color: RGB8,
-    pub duration: f32,
+    /// Duration in milliseconds (max duration ~65.5 seconds)
+    pub duration: u16,
+    /// Color interpolation type
     pub interpolation: Interpolation,
 }
